Addons: Rename texture utils modules.
