+++ examples/jsm/Addons.js
@@ -274,7 +274,7 @@ export * as SceneUtils from './utils/SceneUtils.js';
 export * from './utils/ShadowMapViewer.js';
 export * as SkeletonUtils from './utils/SkeletonUtils.js';
 export * as SortUtils from './utils/SortUtils.js';
-export * from './utils/TextureUtils.js';
+export * from './utils/WebGLTextureUtils.js';
 export * from './utils/UVsDebug.js';
 export * from './utils/WorkerPool.js';
